+++ Assets/MixedRealityToolkit/_Core/Inspectors/MixedRealityInputActionsProfileInspector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2c61f873a01cf6e42a12561d4060c7af
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 8a1b00ef1a7a8b94891466c1f0911027, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
